backend/requirement.txt
Flask
flask-cors
python-dotenv
elevenlabs
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
gTTS==2.3.2
requests==2.28.1
beautifulsoup4==4.11.1

